Add serverDebugMode
This provides initial support for starting the Node.JS server in debug
mode.

Note that triggering the debugger will prevent the server from sending
back replies, which may cause the Bloc UI process to hang waiting for a
response.

Currently it is up to the user to ensure that a deadlock doesn't occur.